Read passphrase from credentials
When configuring the passphrase globally, one of the downsides of the
current implementation is that the pgp plugin must also be installed
globally, which causes issues if you have projects using pre 0.13.5
versions of sbt.

This allows the passphrase to be configured globally without enabling
the pgp plugin globally, by reading the passphrase from the global
credentials.

Since credentials is a task, the task that read it had to be a task, but
pgpPassphrase was a setting. To maintain binary compatibility, rather
than changing phpPassphrase to be a setting, I've introduced a new task
that takes a passphrase defined in the setting in precedence over a
passphrase from credentials.