Use Scala cross building
This makes the plugin use Scala cross building, rather than sbt cross
building, to do cross building. It sets `sbtVersion in pluginCrossBuild`
explicitly based on what the current `scalaVersion` is. So the result is
you can run `+scripted` to cross build against sbt 1.0 and sbt 0.13.

I've also updated the release process to use this, added a note about it
to README, and updated the docs in preparation for a 1.1.0 release.